docker-flask-project/app/requirements.txt
Flask==0.11.1
confluent-kafka==0.11.0
pymysql
flask-sqlalchemy
flask-script
flask-wtf
flask-migrate
py-bcrypt
python-slugify
flask-markdown
-e git+https://github.com/fromzeroedu/flask_uploads#egg=flask_uploads


